+++ src/Ninject.Test/Integration/EnumerableDependenciesTests/WhenServiceRequestsUnconstrainedArrayOfDependencies.cs
@@ -1,5 +1,7 @@
 ﻿namespace Ninject.Tests.Integration.EnumerableDependenciesTests
 {
+    using System.Collections.Generic;
+    using System.Linq;
     using FluentAssertions;
     using Ninject.Tests.Integration.EnumerableDependenciesTests.Fakes;
     using Xunit;
@@ -39,6 +41,24 @@
 
             parent.Should().NotBeNull();
             parent.Children.Count.Should().Be(0);
+        }
+
+        [Fact]
+        public void ArrayIsResolvedIfElementTypeIsExplicitlyBinded()
+        {
+            this.Kernel.Bind<IChild>().To<ChildA>();
+
+            var children = this.Kernel.Get<IChild[]>();
+
+            children.Should().NotBeEmpty();
+        }
+
+        [Fact]
+        public void EmptyArrayIsResolvedIfElementTypeIsMissingBinding()
+        {
+            var children = this.Kernel.Get<IChild[]>();
+
+            children.Should().BeEmpty();
         }
     }
 }