+++ src/Ninject.Test/Integration/EnumerableDependenciesTests/WhenServiceRequestsUnconstrainedListOfConcreteDependencies.cs
@@ -1,5 +1,6 @@
 ﻿namespace Ninject.Tests.Integration.EnumerableDependenciesTests
 {
+    using System.Collections.Generic;
     using FluentAssertions;
     using Ninject.Tests.Integration.EnumerableDependenciesTests.Fakes;
     using Xunit;
@@ -17,7 +18,7 @@
         }
 
         [Fact]
-        public void ServiceIsInjectedWithNonEmptyListIfElementTypeIsExplictlyBinded()
+        public void ServiceIsInjectedWithNonEmptyListIfElementTypeIsExplicitlyBinded()
         {
             this.Kernel.Bind<RequestsListWithConcreteClass>().ToSelf();
             this.Kernel.Bind<ChildA>().ToSelf();
@@ -25,6 +26,24 @@
             var parent = this.Kernel.Get<RequestsListWithConcreteClass>();
 
             parent.Children.Should().NotBeEmpty();
+        }
+
+        [Fact]
+        public void NonEmptyListIsResolvedIfElementTypeIsExplicitlyBinded()
+        {
+            this.Kernel.Bind<ChildA>().ToSelf();
+
+            var children = this.Kernel.Get<IList<ChildA>>();
+
+            children.Should().NotBeEmpty();
+        }
+
+        [Fact]
+        public void EmptyListIsResolvedIfElementTypeIsMissingBinding()
+        {
+            var children = this.Kernel.Get<IList<ChildA>>();
+
+            children.Should().BeEmpty();
         }
     }
 }